+++ crates/ntx-builder/src/state/account.rs
@@ -4,7 +4,77 @@ use miden_node_proto::domain::account::NetworkAccountPrefix;
 use miden_node_proto::domain::note::NetworkNote;
 use miden_objects::account::delta::AccountUpdateDetails;
 use miden_objects::account::{Account, AccountDelta, AccountId};
-use miden_objects::note::Nullifier;
+use miden_objects::block::BlockNumber;
+use miden_objects::note::{Note, Nullifier};
+
+// INFLIGHT NETWORK NOTE
+// ================================================================================================
+
+/// An unconsumed network note that may have failed to execute.
+///
+/// The block number at which the network note was attempted are approximate and may not
+/// reflect the exact block number for which the execution attempt failed. The actual block
+/// will likely be soon after the number that is recorded here.
+#[derive(Debug, Clone)]
+pub struct InflightNetworkNote {
+    note: NetworkNote,
+    attempt_count: usize,
+    last_attempt: Option<BlockNumber>,
+}
+
+impl InflightNetworkNote {
+    /// Creates a new inflight network note.
+    pub fn new(note: NetworkNote) -> Self {
+        Self {
+            note,
+            attempt_count: 0,
+            last_attempt: None,
+        }
+    }
+
+    /// Consumes the inflight network note and returns the inner network note.
+    pub fn into_inner(self) -> NetworkNote {
+        self.note
+    }
+
+    /// Returns a reference to the inner network note.
+    pub fn to_inner(&self) -> &NetworkNote {
+        &self.note
+    }
+
+    /// Returns the number of attempts made to execute the network note.
+    pub fn attempt_count(&self) -> usize {
+        self.attempt_count
+    }
+
+    /// Checks if the network note is available for execution.
+    ///
+    /// The note is available if it can be consumed and the backoff period has passed.
+    pub fn is_available(&self, block_num: BlockNumber) -> bool {
+        let can_consume = self
+            .to_inner()
+            .metadata()
+            .execution_hint()
+            .can_be_consumed(block_num)
+            .unwrap_or(true);
+        can_consume && has_backoff_passed(block_num, self.last_attempt, self.attempt_count)
+    }
+
+    /// Registers a failed attempt to execute the network note at the specified block number.
+    pub fn fail(&mut self, block_num: BlockNumber) {
+        self.last_attempt = Some(block_num);
+        self.attempt_count += 1;
+    }
+}
+
+impl From<InflightNetworkNote> for Note {
+    fn from(value: InflightNetworkNote) -> Self {
+        value.into_inner().into()
+    }
+}
+
+// ACCOUNT STATE
+// ================================================================================================
 
 /// Tracks the state of a network account and its notes.
 pub struct AccountState {
@@ -17,10 +87,10 @@ pub struct AccountState {
     inflight: VecDeque<Account>,
 
     /// Unconsumed notes of this account.
-    available_notes: HashMap<Nullifier, NetworkNote>,
+    available_notes: HashMap<Nullifier, InflightNetworkNote>,
 
     /// Notes which have been consumed by transactions that are still inflight.
-    nullified_notes: HashMap<Nullifier, NetworkNote>,
+    nullified_notes: HashMap<Nullifier, InflightNetworkNote>,
 }
 
 impl AccountState {
@@ -42,6 +112,15 @@ impl AccountState {
             available_notes: HashMap::default(),
             nullified_notes: HashMap::default(),
         }
+    }
+
+    /// Returns an iterator over inflight notes that are not currently within their respective
+    /// backoff periods based on block number.
+    pub fn available_notes(
+        &self,
+        block_num: &BlockNumber,
+    ) -> impl Iterator<Item = &InflightNetworkNote> {
+        self.available_notes.values().filter(|&note| note.is_available(*block_num))
     }
 
     /// Appends a delta to the set of inflight account updates.
@@ -81,7 +160,7 @@ impl AccountState {
 
     /// Adds a new network note making it available for consumption.
     pub fn add_note(&mut self, note: NetworkNote) {
-        self.available_notes.insert(note.nullifier(), note);
+        self.available_notes.insert(note.nullifier(), InflightNetworkNote::new(note));
     }
 
     /// Removes the note completely.
@@ -132,8 +211,9 @@ impl AccountState {
         }
     }
 
-    pub fn notes(&self) -> impl Iterator<Item = &NetworkNote> {
-        self.available_notes.values()
+    /// Drops all notes that have failed to be consumed after a certain number of attempts.
+    pub fn drop_failing_notes(&mut self, max_attempts: usize) {
+        self.available_notes.retain(|_, note| note.attempt_count() < max_attempts);
     }
 
     /// Returns the latest inflight account state.
@@ -153,7 +233,21 @@ impl AccountState {
             && self.available_notes.is_empty()
             && self.nullified_notes.is_empty()
     }
+
+    /// Marks the specified notes as failed.
+    pub fn fail_notes(&mut self, nullifiers: &[Nullifier], block_num: BlockNumber) {
+        for nullifier in nullifiers {
+            if let Some(note) = self.available_notes.get_mut(nullifier) {
+                note.fail(block_num);
+            } else {
+                tracing::warn!(%nullifier, "failed note is not in account's state");
+            }
+        }
+    }
 }
+
+// NETWORK ACCOUNT UPDATE
+// ================================================================================================
 
 #[derive(Clone)]
 pub enum NetworkAccountUpdate {
@@ -182,5 +276,69 @@ impl NetworkAccountUpdate {
             NetworkAccountUpdate::New(account) => account.id(),
             NetworkAccountUpdate::Delta(account_delta) => account_delta.id(),
         }
+    }
+}
+
+// HELPERS
+// ================================================================================================
+
+/// Checks if the backoff block period has passed.
+///
+/// The number of blocks passed since the last attempt must be greater than or equal to
+/// e^(0.25 * `attempt_count`) rounded to the nearest integer.
+///
+/// This evaluates to the following:
+/// - After 1 attempt, the backoff period is 1 block.
+/// - After 3 attempts, the backoff period is 2 blocks.
+/// - After 10 attempts, the backoff period is 12 blocks.
+/// - After 20 attempts, the backoff period is 148 blocks.
+/// - etc...
+#[allow(clippy::cast_precision_loss, clippy::cast_sign_loss)]
+fn has_backoff_passed(
+    chain_tip: BlockNumber,
+    last_attempt: Option<BlockNumber>,
+    attempts: usize,
+) -> bool {
+    if attempts == 0 {
+        return true;
+    }
+    // Compute the number of blocks passed since the last attempt.
+    let blocks_passed = last_attempt
+        .and_then(|last| chain_tip.checked_sub(last.as_u32()))
+        .unwrap_or_default();
+
+    // Compute the exponential backoff threshold: Δ = e^(0.25 * n).
+    let backoff_threshold = (0.25 * attempts as f64).exp().round() as usize;
+
+    // Check if the backoff period has passed.
+    blocks_passed.as_usize() > backoff_threshold
+}
+
+#[cfg(test)]
+mod tests {
+    use miden_objects::block::BlockNumber;
+
+    #[rstest::rstest]
+    #[test]
+    #[case::all_zero(Some(BlockNumber::from(0)), BlockNumber::from(0), 0, true)]
+    #[case::no_attempts(None, BlockNumber::from(0), 0, true)]
+    #[case::one_attempt(Some(BlockNumber::from(0)), BlockNumber::from(2), 1, true)]
+    #[case::three_attempts(Some(BlockNumber::from(0)), BlockNumber::from(3), 3, true)]
+    #[case::ten_attempts(Some(BlockNumber::from(0)), BlockNumber::from(13), 10, true)]
+    #[case::twenty_attempts(Some(BlockNumber::from(0)), BlockNumber::from(149), 20, true)]
+    #[case::one_attempt_false(Some(BlockNumber::from(0)), BlockNumber::from(1), 1, false)]
+    #[case::three_attempts_false(Some(BlockNumber::from(0)), BlockNumber::from(2), 3, false)]
+    #[case::ten_attempts_false(Some(BlockNumber::from(0)), BlockNumber::from(12), 10, false)]
+    #[case::twenty_attempts_false(Some(BlockNumber::from(0)), BlockNumber::from(148), 20, false)]
+    fn backoff_has_passed(
+        #[case] last_attempt_block_num: Option<BlockNumber>,
+        #[case] current_block_num: BlockNumber,
+        #[case] attempt_count: usize,
+        #[case] backoff_should_have_passed: bool,
+    ) {
+        assert_eq!(
+            backoff_should_have_passed,
+            super::has_backoff_passed(current_block_num, last_attempt_block_num, attempt_count)
+        );
     }
 }