+++ crates/ntx-builder/src/state/mod.rs
@@ -2,7 +2,7 @@ use std::collections::hash_map::Entry;
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet, VecDeque};
 use std::num::NonZeroUsize;
 
-use account::{AccountState, NetworkAccountUpdate};
+use account::{AccountState, InflightNetworkNote, NetworkAccountUpdate};
 use anyhow::Context;
 use miden_node_proto::domain::account::NetworkAccountPrefix;
 use miden_node_proto::domain::mempool::MempoolEvent;
@@ -11,7 +11,7 @@ use miden_node_utils::tracing::OpenTelemetrySpanExt;
 use miden_objects::account::Account;
 use miden_objects::account::delta::AccountUpdateDetails;
 use miden_objects::block::{BlockHeader, BlockNumber};
-use miden_objects::note::Nullifier;
+use miden_objects::note::{Note, Nullifier};
 use miden_objects::transaction::{PartialBlockchain, TransactionId};
 use tracing::instrument;
 
@@ -30,12 +30,13 @@ const MAX_BLOCK_COUNT: usize = 4;
 ///
 /// Contains the data pertaining to a specific network account which can be used to build a network
 /// transaction.
+#[derive(Clone)]
 pub struct TransactionCandidate {
     /// The current inflight state of the account.
     pub account: Account,
 
     /// A set of notes addressed to this network account.
-    pub notes: Vec<NetworkNote>,
+    pub notes: Vec<InflightNetworkNote>,
 
     /// The latest locally committed block header.
     ///
@@ -88,6 +89,9 @@ pub struct State {
 }
 
 impl State {
+    /// Maximum number of attempts to execute a network note.
+    const MAX_NOTE_ATTEMPTS: usize = 30;
+
     /// Load's all available network notes from the store, along with the required account states.
     #[instrument(target = COMPONENT, name = "ntx.state.load", skip_all)]
     pub async fn load(store: StoreClient) -> Result<Self, StoreError> {
@@ -152,10 +156,12 @@ impl State {
                 continue;
             }
 
-            let account = self.accounts.get(&candidate).expect("queue account must be tracked");
+            let account = self.accounts.get_mut(&candidate).expect("queue account must be tracked");
+
+            // Remove notes that have failed too many times.
+            account.drop_failing_notes(Self::MAX_NOTE_ATTEMPTS);
 
             // Skip empty accounts, and prune them.
-            //
             // This is how we keep the number of accounts bounded.
             if account.is_empty() {
                 // We don't need to prune the inflight transactions because if the account is empty,
@@ -166,7 +172,12 @@ impl State {
                 continue;
             }
 
-            let notes = account.notes().take(limit.get()).cloned().collect::<Vec<_>>();
+            // Select notes from the account that can be consumed or are ready for a retry.
+            let notes = account
+                .available_notes(&self.chain_tip_header.block_num())
+                .take(limit.get())
+                .cloned()
+                .collect::<Vec<_>>();
 
             // Skip accounts with no available notes.
             if notes.is_empty() {
@@ -208,11 +219,32 @@ impl State {
         self.chain_mmr.prune_to(..pruned_block_height.into());
     }
 
+    /// Marks notes of a previously selected candidate as failed.
+    ///
+    /// Does not remove the candidate from the in-progress pool.
+    #[instrument(target = COMPONENT, name = "ntx.state.notes_failed", skip_all)]
+    pub fn notes_failed(
+        &mut self,
+        candidate: NetworkAccountPrefix,
+        notes: &[Note],
+        block_num: BlockNumber,
+    ) {
+        if let Some(account) = self.accounts.get_mut(&candidate) {
+            let nullifiers = notes.iter().map(Note::nullifier).collect::<Vec<_>>();
+            account.fail_notes(nullifiers.as_slice(), block_num);
+        } else {
+            tracing::error!(account.prefix=%candidate, "failed network notes have no local account state");
+        }
+    }
+
     /// Marks a previously selected candidate account as failed, allowing it to be available for
     /// selection again.
+    ///
+    /// All notes in the candidate will be marked as failed.
     #[instrument(target = COMPONENT, name = "ntx.state.candidate_failed", skip_all)]
     pub fn candidate_failed(&mut self, candidate: NetworkAccountPrefix) {
         self.in_progress.remove(&candidate);
+
         self.inject_telemetry();
     }